requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
openai>=1.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
jsonlines>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Grounding DINO dependencies
addict>=2.4.0
yapf>=0.40.2
supervision>=0.16.0